feat(expo): migrate FollowedListsModal to /subscribed-lists formSheet
Replace the RN Modal with a router formSheet so the subscribed-lists sheet
uses the same native header + grabber language as Event Details. Tapping a
list dismisses the sheet before pushing the list detail so the user does not
land back on subscribed-lists after closing the list screen.
