+++ apps/expo/src/app/(modals)/subscribed-lists.tsx
@@ -1,0 +1,159 @@
+import React, { useCallback } from "react";
+import {
+  ActivityIndicator,
+  FlatList,
+  Platform,
+  Share,
+  Text,
+  TouchableOpacity,
+  View,
+} from "react-native";
+import { router, Stack } from "expo-router";
+import { useMutation, useQuery } from "convex/react";
+
+import type { Doc } from "@soonlist/backend/convex/_generated/dataModel";
+import { api } from "@soonlist/backend/convex/_generated/api";
+
+import { List, ShareIcon } from "~/components/icons";
+import { SubscribeButton } from "~/components/SubscribeButton";
+import Config from "~/utils/config";
+import { logError } from "~/utils/errorLogging";
+import { toast } from "~/utils/feedback";
+
+export default function SubscribedListsScreen() {
+  const followedLists = useQuery(api.lists.getFollowedLists);
+  const unfollowListMutation = useMutation(
+    api.lists.unfollowList,
+  ).withOptimisticUpdate((localStore, args) => {
+    const current = localStore.getQuery(api.lists.getFollowedLists, {});
+    if (current === undefined) return;
+    localStore.setQuery(
+      api.lists.getFollowedLists,
+      {},
+      current.filter((l) => l.id !== args.listId),
+    );
+  });
+
+  const handleUnfollow = useCallback(
+    (listId: string) => {
+      unfollowListMutation({ listId }).catch((error: unknown) => {
+        logError("Error unfollowing list", error);
+        toast.error("Failed to unsubscribe");
+      });
+    },
+    [unfollowListMutation],
+  );
+
+  const handleShareList = useCallback(
+    async (listName: string, listSlug?: string) => {
+      const shareUrl = listSlug
+        ? `${Config.apiBaseUrl}/list/${listSlug}`
+        : Config.apiBaseUrl;
+      try {
+        await Share.share(
+          Platform.OS === "ios"
+            ? { url: shareUrl }
+            : { message: `Check out ${listName} on Soonlist: ${shareUrl}` },
+        );
+      } catch (error) {
+        logError("Error sharing list", error);
+      }
+    },
+    [],
+  );
+
+  const handleListPress = useCallback((list: Doc<"lists">) => {
+    if (!list.slug) return;
+    // Dismiss the sheet first so the user doesn't land back on subscribed
+    // lists after closing the list detail.
+    router.back();
+    router.push(`/list/${list.slug}`);
+  }, []);
+
+  return (
+    <View className="flex-1 bg-white">
+      <Stack.Screen
+        options={{
+          title: "Subscribed lists",
+          headerRight: () => (
+            <TouchableOpacity onPress={() => router.back()} activeOpacity={0.7}>
+              <Text className="text-base font-semibold text-interactive-1">
+                Done
+              </Text>
+            </TouchableOpacity>
+          ),
+        }}
+      />
+
+      {followedLists === undefined ? (
+        <View className="flex-1 items-center justify-center">
+          <ActivityIndicator size="large" color="#5A32FB" />
+        </View>
+      ) : followedLists.length === 0 ? (
+        <View className="flex-1 items-center justify-center px-4">
+          <Text className="text-base text-neutral-2">
+            Not subscribed to any lists yet
+          </Text>
+        </View>
+      ) : (
+        <FlatList
+          data={followedLists}
+          keyExtractor={(list) => list.id}
+          contentContainerStyle={{ paddingHorizontal: 16, paddingBottom: 16 }}
+          renderItem={({ item: list, index }) => {
+            const isFirst = index === 0;
+            const isLast = index === followedLists.length - 1;
+            return (
+              <View
+                className={`flex-row items-center bg-interactive-3/60 px-4 py-2.5 ${
+                  isFirst ? "rounded-t-2xl pt-3" : ""
+                } ${isLast ? "rounded-b-2xl pb-3" : ""}`}
+              >
+                <TouchableOpacity
+                  className="min-w-0 flex-1 flex-row items-center"
+                  onPress={() => handleListPress(list)}
+                  activeOpacity={0.7}
+                  accessibilityRole="button"
+                  accessibilityLabel={`Open list ${list.name}`}
+                >
+                  <View className="h-11 w-11 shrink-0 items-center justify-center rounded-full bg-interactive-3">
+                    <List size={20} color="#5A32FB" />
+                  </View>
+                  <View className="ml-3 min-w-0 flex-1">
+                    <Text
+                      className="text-base font-semibold text-neutral-1"
+                      numberOfLines={1}
+                    >
+                      {list.name}
+                    </Text>
+                  </View>
+                </TouchableOpacity>
+
+                <TouchableOpacity
+                  onPress={() =>
+                    void handleShareList(list.name, list.slug ?? undefined)
+                  }
+                  className="ml-2 rounded-full p-2"
+                  activeOpacity={0.7}
+                  accessibilityLabel={`Share ${list.name}`}
+                  hitSlop={{ top: 8, bottom: 8, left: 8, right: 8 }}
+                >
+                  <ShareIcon size={18} color="#5A32FB" />
+                </TouchableOpacity>
+
+                <View className="ml-1">
+                  <SubscribeButton
+                    isSubscribed={true}
+                    onPress={() => handleUnfollow(list.id)}
+                    size="sm"
+                    accessibilityLabel={`Unsubscribe from ${list.name}`}
+                  />
+                </View>
+              </View>
+            );
+          }}
+        />
+      )}
+    </View>
+  );
+}
